Move BlogPostHeaderTagItem to shared directory
related: #30
